move filter components into subdirectories
