+++ frontend/src/app/filter/boolean-filter/boolean-filter.component.html
