+++ frontend/src/app/filter/boolean-filter/boolean-filter.component.spec.ts
@@ -1,8 +1,8 @@
 import { ComponentFixture, TestBed, fakeAsync, tick, waitForAsync } from '@angular/core/testing';
-import { mockCorpus3, mockField } from '../../mock-data/corpus';
+import { mockCorpus3, mockField } from '../../../mock-data/corpus';
 
-import { commonTestBed } from '../common-test-bed';
-import { BooleanFilter, QueryModel } from '../models';
+import { commonTestBed } from '../../common-test-bed';
+import { BooleanFilter, QueryModel } from '../../models';
 import { BooleanFilterComponent } from './boolean-filter.component';
 import { By } from '@angular/platform-browser';
