+++ frontend/src/app/filter/date-filter/date-filter.component.scss
@@ -1,4 +1,4 @@
-@import "../../_utilities";
+@import "../../../_utilities";
 
 ::ng-deep .p-datepicker {
     margin: 1px 0;
