+++ frontend/src/app/filter/date-filter/date-filter.component.spec.ts
@@ -1,8 +1,8 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { mockCorpus3, mockFieldDate } from '../../mock-data/corpus';
+import { mockCorpus3, mockFieldDate } from '../../../mock-data/corpus';
 
-import { commonTestBed } from '../common-test-bed';
-import { DateFilter, DateFilterData, QueryModel } from '../models';
+import { commonTestBed } from '../../common-test-bed';
+import { DateFilter, DateFilterData, QueryModel } from '../../models';
 
 import { DateFilterComponent } from './date-filter.component';
