+++ frontend/src/app/filter/date-filter/date-filter.component.ts
@@ -1,14 +1,14 @@
 import { Component } from '@angular/core';
 import * as _ from 'lodash';
 
-import { DateFilter } from '../models';
-import { BaseFilterComponent } from './base-filter.component';
+import { DateFilter } from '../../models';
+import { BaseFilterComponent } from '../base-filter.component';
 import { BehaviorSubject, combineLatest } from 'rxjs';
 
 @Component({
-  selector: 'ia-date-filter',
-  templateUrl: './date-filter.component.html',
-  styleUrls: ['./date-filter.component.scss']
+    selector: 'ia-date-filter',
+    templateUrl: './date-filter.component.html',
+    styleUrls: ['./date-filter.component.scss']
 })
 export class DateFilterComponent extends BaseFilterComponent<DateFilter> {
     public minDate: Date;
@@ -25,7 +25,7 @@ export class DateFilterComponent extends BaseFilterComponent<DateFilter> {
         this.selectedMaxDate = new BehaviorSubject(filter.currentData.max);
 
         combineLatest([this.selectedMinDate, this.selectedMaxDate]).subscribe(([min, max]) =>
-            this.update({min, max})
+            this.update({ min, max })
         );
     }
