+++ frontend/src/app/filter/index.ts
@@ -1,6 +1,6 @@
 export * from './base-filter.component';
-export * from './boolean-filter.component';
-export * from './date-filter.component';
+export * from './boolean-filter/boolean-filter.component';
+export * from './date-filter/date-filter.component';
 export * from './filter-manager.component';
-export * from './multiple-choice-filter.component';
-export * from './range-filter.component';
+export * from './multiple-choice-filter/multiple-choice-filter.component';
+export * from './range-filter/range-filter.component';
