+++ frontend/src/app/filter/multiple-choice-filter/multiple-choice-filter.component.scss
@@ -1,4 +1,4 @@
-@import "../../_utilities";
+@import "../../../_utilities";
 
 ::ng-deep .p-multiselect-panel {
     max-width: 300px;
